added dataBinding for ListFragment

+++ app/src/main/java/com/example/zach/weatherapp/data/Forecast.kt
@@ -1,7 +1,7 @@
 package com.example.zach.weatherapp.data
 
 data class Forecast(var temp:Double,
-                    var pressure:Int,
+                    var pressure:Double,
                     var humidity: Int,
                     var temp_min: Double,
                     var temp_max: Double)
